fix(GCC13): -Wshadow / -Wsign-compare / -Wformat-truncation on Ubuntu 24.04
GCC 13's -Werror (CI on ubuntu-24.04) is stricter than the existing
clang/MSVC matrix. Three classes of issue surfaced when a fresh clone
of main built on a stock /usr/bin/g++:

  1. src/platform/time.cpp — snprintf("%04d-%02d-%02d", tm.tm_year+1900,
     ...) produced -Wformat-truncation because tm_year is wide-typed.
     Clamp the inputs into known small ranges first; the formatted
     output is provably <= 10 chars into char[16].

  2. src/engine/table.cpp — `off + 4 <= rec.size()` mixed signed
     uint16_t with size_t. Cast to size_t.

  3. src/abi/ace_exports.cpp — 13 lambdas captured an outer `s` (the
     state() reference) through their `std::string s` param,
     triggering -Wshadow. Rename each param to `sl` (and one `m` ->
     `mo` in the Julian-day DATEDIFF helper). No behavioral change.

345 -> 346 unit tests still all passing on Windows MSVC.

+++ src/engine/table.cpp
@@ -351,7 +351,8 @@ util::Result<void> Table::append_record() {
         if (!bumped) continue;   // tolerate non-autoinc-capable drivers
         std::uint32_t v = bumped.value();
         std::uint16_t off = fs[i].record_offset;
-        if (fs[i].length >= 4 && off + 4 <= rec.size()) {
+        if (fs[i].length >= 4 &&
+            static_cast<std::size_t>(off) + 4 <= rec.size()) {
             rec[off    ] = static_cast<std::uint8_t>( v        & 0xFFu);
             rec[off + 1] = static_cast<std::uint8_t>((v >>  8) & 0xFFu);
             rec[off + 2] = static_cast<std::uint8_t>((v >> 16) & 0xFFu);
